Refactors domain objects to use IDs and BigDecimal
Updates domain objects CoffeeShop, Comment and User
to have explicit ID fields to represent them in the DB.

Converts the rating field in the Comment class and Visit calculation
to use BigDecimal for improved precision.

Exposes an unmodifiable list of comments in the Visit class and
introduces an addComment method to handle adding comments to a visit.